+++ fiddles/angular2-cli/fiddle-0012-PanelReOrdering/e2e/app.e2e-spec.ts
@@ -1,0 +1,14 @@
+import { PanelReOrderingPage } from './app.po';
+
+describe('panel-re-ordering App', () => {
+  let page: PanelReOrderingPage;
+
+  beforeEach(() => {
+    page = new PanelReOrderingPage();
+  });
+
+  it('should display welcome message', () => {
+    page.navigateTo();
+    expect(page.getParagraphText()).toEqual('Welcome to app!');
+  });
+});